requirements.txt
acme==1.12.0
Brlapi==0.8.2
certbot==1.12.0
certbot-nginx==1.10.1
certifi==2020.6.20
chardet==4.0.0
ConfigArgParse==1.2.3
configobj==5.0.6
cryptography==3.3.2
distro==1.5.0
et-xmlfile==1.1.0
greenlet==1.1.2
httplib2==0.18.1
idna==2.10
josepy==1.2.0
louis==3.16.0
numpy==1.23.1
openpyxl==3.0.10
pandas==1.4.3
parsedatetime==2.6
psycopg2-binary==2.9.3
pycairo==1.16.2
pycurl==7.43.0.6
PyGObject==3.38.0
PyICU==2.5
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyRFC3339==1.1
PySimpleSOAP==1.16.2
python-apt==2.2.1
python-dateutil==2.8.2
python-debian==0.1.39
python-debianbts==3.1.0
pytz==2021.1
pyxdg==0.27
reportbug==7.10.3+deb11u1
requests==2.25.1
requests-toolbelt==0.9.1
six==1.16.0
SQLAlchemy==1.4.39
urllib3==1.26.5
xdg==5
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.0.1
zope.interface==5.2.0
